Toolbar: break up props stories
