+++ src/layout/Toolbar/props/props.story.tsx
@@ -1,0 +1,9 @@
+import { Toolbar } from '../Toolbar';
+
+export default {
+  title: 'layout/Toolbar/props',
+  component: Toolbar,
+};
+
+export { Attach } from './attach';
+export { State } from './state';
